Query history improvments (#336)
* Fix sort issues on the QueryHistory

* What changed
1. QueryStoreGridControl.Selection.cs — ViewHistory_Click now walks up the visual tree to find the parent QuerySessionControl and calls AddHistorySubTab(), placing the history as a sub-tab alongside "Query Store — DB" and "QS Overview" (instead of at the top-level MainTabControl).
2. QuerySessionControl.QueryStore.cs — Added two new methods:
• AddHistorySubTab() — creates a closeable sub-tab with the history control, including long-press (~500ms) to detach into a free-floating window
• DetachHistorySubTabToWindow() — pops the history content into a standalone non-modal window (movable to another screen). When that window is minimized or closed, the content automatically re-docks back as a sub-tab
3. MainWindow.Tabs.cs — Removed the now-unused AddHistoryTab() method since history tabs no longer live at the top level

User flow
• View History → opens as a sub-tab next to Query Store / QS Overview
• Long-press the history sub-tab → detaches into a free window (can move to another screen)
• Minimize or close the free window → content returns to the sub-tab

* What was added
1. QueryStoreService.FetchPlanByHashAsync() — New method that fetches a plan XML from sys.query_store_plan by query_plan_hash. The oldest parameter controls whether it returns the first (smallest plan_id) or last (largest plan_id) plan.
2. Context menu on QueryStoreHistoryControl — Right-clicking on the DataGrid or chart shows:
• "Load the First Plan" — fetches the oldest plan for the selected plan hash
• "Load the Last Plan" — fetches the most recent plan for the selected plan hash
The selected plan hash is determined from either the grid selection or chart selection.
3. PlanLoadRequested event — Raised when the user picks a plan from the context menu. The parent QuerySessionControl subscribes to this event and opens the plan XML as a new sub-tab (using the existing AddPlanTab mechanism).

User flow
• Select a row in the history grid (or a dot on the chart)
• Right-click → "Load the First Plan" or "Load the Last Plan"
• The plan opens as a new sub-tab (e.g., "QS 42 / 17")

* Code review :

Bugs Fixed
1. Shared ContextMenu — BuildContextMenu() now creates separate ContextMenu instances for DataGrid and Chart via CreatePlanContextMenu()
2. Silent catch — LoadPlanFromSelection now shows errors in StatusText with catch (Exception ex)
3. Wrong type check — Close_Click now checks is not PlanViewer.App.MainWindow instead of IClassicDesktopStyleApplicationLifetime
4. Event handler leak — AddHistorySubTab now unsubscribes before subscribing: -= OnHistoryPlanLoadRequested before +=
Improvements
5. Loading feedback — Shows "Loading plan…" / "Plan not found" / error in StatusText
6. Disable menu when no selection — Opening handler disables items when GetSelectedPlanHash() is null
7. IndexOf — Added clarifying comment (list is <500 items, no better API available)
8. ScrollIntoView — Moved after ItemsSource reset so the target row exists
9. Long-press duplication — Extracted TabHeaderLongPressBehavior helper, used in both MainWindow.Tabs.cs and QuerySessionControl.QueryStore.cs
10. HistoryPlanLoadEventArgs — Moved to its own file
Nits
11. Removed unnecessary ?. on non-nullable orderBy parameter
12. Changed bare catch to catch (Exception) in tooltip handler
13. Changed bare catch to catch (Exception ex) in LoadPlanFromSelection

* When the MainWindow closes, it now iterates all open windows via IClassicDesktopStyleApplicationLifetime.Windows and closes any that aren't the MainWindow itself. This ensures all detached free-floating windows (from tab detach, history detach, advice windows, etc.) are closed when the app's main window is closed.

* Must-fix fixed

* use pin/unpin button to fix ux-must-fix

* fix error truncation to 80 chars : truncation removed

* fix most remaining nice to have

* add spinner and cancel for history fetches
